+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # NEXT RELEASE
 
+### v0.69.21
+* Fix filtering of user's own toots out of their timeline
+* Return more `MastoApi` properties in `getCurrentState()`
+
 ### v0.69.20
 * Add `allEmojis()` getter to `Toot`
 * (Demo App) Larger font for version information